First Commit
Resolved issues with incorrect and mismatched values for outlet quality in the different test cases. Began creation of a C++ Pipe class for future development. Next steps are to continue to serialize Matlab code, program in the three optimization cases, use a more complete version of the pressure drop calculations (to include acceleration effects), and develop some nice looking graphs. Start by testing a straight pipe with no heat input to ensure that the model is correct, then increase in complexity