Added summation for total pressure drop through pipe
Added summation for finding the total pressure drop through the pipe for each case. This is using the same equation for finding each pressure increment as Nick Larimer used. I still need to add the terms for acceleration, compressibility, and flashing, but I wanted to make this commit so that I could go back and get the numbers for the original equation (Nick's version) if needed. It is worth noting that the numbers I get for "pressure_inc", "pressure_max", and "pressure_min" are very similar, but not exactly equal to, the numbers obtained for the corresponding variables in Nick's code. I am not sure yet if it is worth investigating the discrepancy, as the numbers are very close. In addition, I will be changing the pressure drop equation soon anyway.