Added output normalizations and flow quality graph
Added normalization of each output parameter, optimization case calculations, and graphs for flow quality as a function of axial distance. This graph appears to be correct for the three optimization cases. *Note: I did not commit changes for utilizing a 5D array for incremental pressure drops (so that I can access the data for each angle for graphing purposes later on)