Added graphs
Added 6 graphs: vapor quality as a function of axial distance (for the three optimization cases), total pressure drop as a function of pipe angle (linear and log, only for utopia point), incremental pressure drop as a function of axial distance (for the three optimization cases, both linear and log), and cumulative pressure drop as a function of axial distance (all three cases, linear and log). All plots seem to agree with traditional equations and intuition, Dr. Rau seems to agree. Next step is to investigate the utopia angle undergoing a sudden drop off to zero degrees near 900 kW.