Added file for plots only
I thought I committed and pushed this file last week but apparently not. Plots.m is used for all plots/graphs from now on. This will help block out my code more (in addition to creating and using functions). A separate section in the main file could be used instead, but that would just clutter up that file more than necessary. The plots can be run and changed quickly and conveniently once all the necessary variables are in the workspace (so the main code need only be run once). I also made a few minor changes to the main code, one being the creation of a multidimensional array for storing incremental heat transfer coefficients (for plotting purposes).