Made some minor changes to master file and plot file, created classes for HEM model and Saturated Water
Major changes include fixing a few equations (void fraction), updating absolute pressure, and adding capability to test subcooled fluids. This is done through updating the temperature in the correct manner if the inlet temperature is lower than the saturation temperature. There were some issues with the flashing and compressibility factors (which are also problems with the derivatives of specific volume and enthalpy with respect to pressure), so they are currently being calculated, but not used in the pressure drop eequations. A fix for that will be coming in the next commit/version. In addition, the beginnings of classes for SaturatedWater and the HEM model were created. The structuree is there, but some more equations need to be added in. In addition, the same issues with the volume and enthalpy derivatives mentioned above need fixed in the Saturated Water class file. These classes are meant to be independent of the geometry of the domain in question.