Restructured code and file structure
I am now using an object-oriented approach for my thesis work. The current version of the code is on /OOPCode, and the rest of the code is stored in the other folders. Avoid trying to use or understand the older code, I am no longer updating it. Only reference it if you need to find where older results came from. I still need to comment up my current version of the code.